Converted most UniversalComponents to in-file SC
