+++ src/components/ActivityLinks/ActivityLinks.jsx
@@ -8,11 +8,11 @@ import * as U from '../../components/TuxComponents/UniversalComponents';
 const LinkList = ({ links }) => {
   <ul>
     {links.map((link, index) => (
-      <U.ListLinks>
-        <U.LinkTo noDecor key={index} href={link.url}>
+      <ListLinks>
+        <LinkTo noDecor key={index} href={link.url}>
           {link.text}
-        </U.LinkTo>
-      </U.ListLinks>
+        </LinkTo>
+      </ListLinks>
     ))}
   </ul>;
 };
@@ -21,13 +21,13 @@ const ActivityLinks = () => {
   const { links } = useContext(SessionContext);
   return (
     <U.Sub6ColGrid>
-      <U.Heading2>More Heuristics Resources</U.Heading2>
+      <Heading1>More Heuristics Resources</Heading1>
       <FlexBox>
       <Text>
         Don't stop here! There are a lot more ways to learn about accessibility
         and many websites designed to help you better design for accessibility.
       </Text>
-      <Heading6>Articles:</Heading6>
+      <Heading2>Articles:</Heading2>
       <ul>
         {links.map((link, index) => (
           <ListLinks>
@@ -49,18 +49,6 @@ const FlexBox = styled.div`
   margin: 40px
 `;
 
-const Text = styled.p`
-  font: var(--pop-reg);
-  line-height: 20px;
-  font-size: 24px;
-`;
-
-const Heading6 = styled.h6`
-  font: var(--table);
-  font-weight: 700;
-  line-height: 24px;
-`;
-
 const ListLinks = styled.li`
   font: var(--table);
   line-height: 24px;
@@ -68,7 +56,28 @@ const ListLinks = styled.li`
   color: var(--present);
 `;
 
+const Heading1 = styled.h1`
+  font: 700 48px 'Poppins', sans-serif;
+  line-height: 24px;
+  color: var(--extraLg-text);
+  text-align: center;
+  margin-top: 75px;
+`;
+
+const Text = styled.p`
+  font: var(--pop-reg);
+  line-height: 20px;
+  font-size: 24px;
+`;
+
+const Heading2 = styled.h2`
+  font: var(--table);
+  font-weight: 700;
+  line-height: 24px;
+`;
+
 const LinkTo = styled(Link)`
   color: var(--link-text);
   list-style: none;
 `;
+
