+++ src/components/GroupUserTable/GroupUserTable.jsx
@@ -1,31 +1,77 @@
 import React from 'react';
-import * as U from '../TuxComponents/UniversalComponents';
+import styled, { css } from 'styled-components';
 
 const GroupUserTable = ({groups, selectedGroupIndex}) => {
   const selectedGroup = groups[selectedGroupIndex]
   const {invited} = selectedGroup
-  return ( <U.Table>
+  return ( <Table>
     <thead>
-    <U.TableRow> 
-        <U.TableHeader left colspan="1">Student Name</U.TableHeader>
-        <U.TableHeader left colspan="1">Email</U.TableHeader>
-        <U.TableHeader colspan="1">Delete</U.TableHeader>
-    </U.TableRow>
+    <TableRow> 
+        <TableHeader left colspan="1">Student Name</TableHeader>
+        <TableHeader left colspan="1">Email</TableHeader>
+        <TableHeader colspan="1">Delete</TableHeader>
+    </TableRow>
     </thead>
     <tbody>
       {invited.length > 0 ? (
         invited.map((user, index) => (
-          <U.TableRow>
-            <U.TableData>{user.name && user.name.length > 0 ? user.name : '--- not sign up ---'}</U.TableData>
-            <U.TableData>{user.email}</U.TableData>
-            <U.TableData left><U.Icon25 src="/images/icons/Trashcan.png"></U.Icon25></U.TableData>
-          </U.TableRow>
+          <TableRow>
+            <TableData>{user.name && user.name.length > 0 ? user.name : '--- not sign up ---'}</TableData>
+            <TableData>{user.email}</TableData>
+            <TableData left><Icon25 src="/images/icons/Trashcan.png"></Icon25></TableData>
+          </TableRow>
         ))
       ) : (
         <>No users yet</>
       )}
     </tbody>
-  </U.Table> );
+  </Table> );
 }
  
 export default GroupUserTable;
+
+
+/*---------- Tables ----------*/
+const Table = styled.table`
+  width: 100%;
+`;
+
+const TableRow = styled.tr`
+  border-bottom: 1px solid #cccccc;
+`;
+
+const TableHeader = styled.th`
+  font: var(--table);
+  font-weight: 600;
+  line-height: 24px;
+  color: rgba(51, 51, 51, 0.75);
+  height: 50px;
+  text-align: left;
+
+  ${(props) =>
+    props.left &&
+    css`
+      padding-left: 20px;
+    `}
+`;
+
+const TableData = styled.td`
+  font: var(--table);
+  color: rgba(51, 51, 51, 0.75);
+  font-weight: 400;
+  height: 50px;
+  text-align: left;
+
+  ${(props) =>
+    props.left &&
+    css`
+      padding-left: 15px;
+    `}
+`;
+
+
+/*---------- Images ----------*/
+const Icon25 = styled.img`
+  width: 25px;
+  height: 25px;
+`;
