+++ src/components/SignupForm/EmailForm.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
+import {Link} from 'react-router-dom';
 import * as U from '../TuxComponents/UniversalComponents';
 import * as L from '../TuxComponents/LoginComponents';
 import styled, { css } from 'styled-components';
+import { Button280 } from '../TuxComponents/elements'
 
 const EmailForm = ({ errors, values, handleChange, touched }) => {
   const { userData } = values
@@ -11,7 +13,7 @@ const EmailForm = ({ errors, values, handleChange, touched }) => {
         <U.FlexBox realtive>
           <L.TuxFlower src="/images/TuxFlower.svg" alt="tux logo, a blue, yellow and red flower"></L.TuxFlower>
           <L.LoginBox flexStart>
-            <U.Heading1 margin50>Enter an Email</U.Heading1>
+            <Heading1 margin50>Enter an Email</Heading1>
             <U.Normal setup>
               We'll use this email to set up your account.
             </U.Normal>
@@ -27,8 +29,8 @@ const EmailForm = ({ errors, values, handleChange, touched }) => {
             ></SmallInput>
             {errors.email && <div>{errors.email}</div>}
 
-            <U.WideBtn
-              extraWide
+            <Button280
+              wide350
               disabled={
                 !errors.email && userData.email && userData.email.length > 0
                   ? false
@@ -40,13 +42,13 @@ const EmailForm = ({ errors, values, handleChange, touched }) => {
               onClick={handleChange}
             >
               Continue
-            </U.WideBtn>
+            </Button280>
           </L.LoginBox>
         </U.FlexBox>
       </U.FlexBox>
       <U.FlexBox login>
         <U.Normal alignRight>Have an Account?&nbsp;</U.Normal>
-        <U.LinkTo to="/login">Log In</U.LinkTo>
+        <LinkTo to="/login">Log In</LinkTo>
       </U.FlexBox>
     </L.Main>
   );
@@ -76,4 +78,20 @@ export const SmallInput = styled.input`
     css`
       margin: 98px 10px 88px 0;
     `}
+`;
+
+export const LinkTo = styled(Link)`
+  color: var(--link-text);
+
+  ${(props) => props.noDecor && css `
+    list-style: none;
+
+  `}
+`;
+
+const Heading1 = styled.h1`
+  font-size: 36px;
+  line-height: 54px;
+  font-weight: 500;
+  margin-top: 50px;
 `;
